refactor: extract decode attn wave validation
